Update patch set 1
Patch Set 1:

> I cherry-picked this to boron. Does it also apply to Beryllium? I
 > think so, but I'm not sure.

The patch that initially documented the cluster-admin API was not cherry-picked to Beryllium. The cluster-admin.yang file and the API as present in Beryllium too, but it is part of a different artifact, so paths need adjustment, and I guess it would be nice if someone tested it too.

Patch-set: 1
Reviewer: Gerrit User 491 <491@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0