Updated git submodules
Project: integration/test master a886e8900da363152375f985d525911af809c9be

Configure timeout in Open_Connection_* keywords

The value of this timeout argument is passed "as is" to the
"Open Connection" keyword invocation that is inside of these
keywords. This is needed by suites that require a different
timeout from the default one.

Change-Id: I6681343be3bfbb43e198e347814a16a4cde9d19e
Signed-off-by: Jozef Behran <jbehran@cisco.com>
Signed-off-by: Vratko Polak <vrpolak@cisco.com>