Updated git submodules
Project: genius master ffd67c3748872793984a31e70345dea168ca6be2

Merge "mdsalutil-api datastoreutils thread pools etc. shutdown on bundle stop"

mdsalutil-api datastoreutils thread pools etc. shutdown on bundle stop

using an OSGi Activator, which isn't how I would do it in an ideal
world, but for the current design this is probably the best we can do;
the clean solution is as proposed in
https://git.opendaylight.org/gerrit/#/c/51431/14/jobcoordinator/jobcoordinator-impl/src/main/java/org/opendaylight/infrautils/jobcoordinator/internal/JobCoordinatorImpl.java

originally built in the hope that it would solve a problem with the
SingleFeatureTest (SFT) under Karaf 4, which it didn't - but this still
seems like a good idea, for possible future support of clean and fast
shut down.

probably will solve Bug 6564 in NetvirtIT (that wasn't the original
motivation, but as a "side effect" what's going on there should not
happen anymore)

Change-Id: I3e5e40ce5d23a15369732f79d0e6b0153a0cc351
Signed-off-by: Michael Vorburger <vorburger@redhat.com>