Updated git submodules
Project: netvirt stable/boron 99a36e584168ac8bbf9e2b05925a16809a6bd82e

Bug 7939 - Remote flows missing in Table 21

Due to racing between createRemoteFibEntry() invocations
from multiple DJCs of different keys within VRFEntryListener,
when two threads try to create the same remote flow, it lands
up as OptimisticLockFailedException and that results in none
of these threads write showing up in the flow-inventory.

With this patch we donot eliminate the race, but rather attempt
to reduce the invocations to createRemoteFibEntry there reducing
the occurrence of the problem.

Changing Keys in VRFEntryListener requires more careful thought
as these keys not only protect DS writes but also bring in
batching at scale.   And so that part has to be addressed in a
different review.

Change-Id: I7bf80742980002b471520fc3e50ed1c28baa1945
Signed-off-by: Vivekanandan Narasimhan <n.vivekanandan@ericsson.com>