Updated git submodules
Project: genius master 97d65bfbd295d4d2ff3da2e504df9f0e2f773805

LockManager unlock should not pretend its successful when it was not

or was this intentionally done as-is, for some reason I'm missing?

It's seems strange to "hide" error during unlock (currently it logs an
INFO not even an ERROR, and returns sucess, when if there was a
problem).

If so, then it should be documented with an in-line comment...

Change-Id: I977d1cba241c735ba97325fece4ec09964a5355f
Signed-off-by: Michael Vorburger <vorburger@redhat.com>