Update patch set 3
Patch Set 3: Code-Review-1

(2 comments)

Mostly because of the incorrect ordering of using a virtualenv. Personally I'd rather not documented virtualenv as it is possible to use lftools without one and the user should already know how to setup virtualenv if they wish to use one. Otherwise we might as well document how to install and setup a virtualenv as well which is out of the scope of this doc IMO.

Patch-set: 3
Label: Code-Review=-1