Updated git submodules
Project: ovsdb stable/carbon 4b0913ac93fe903e5400bcc1d1cf7abde6e44155

bug 8257 handling back to back mcast mac updates

Making the remove mcast mac participate in depenency resolution
flow.

Put the delete event on hold while the previous add/update on
the same key is still in transit.

Delete mcast mac event resumes after we receive the
 response for the previous create.

Mark the state of the mcast mac key in transit while we execute its delete/add/update.

Change-Id: I0b15d82dad346bd2c35c0368b4d7f63dc3bfa09d
Signed-off-by: K.V Suneelu Verma <k.v.suneelu.verma@ericsson.com>
(cherry picked from commit 1c48d4d9765a0fabc1abce8e90ab263580a40870)


Bump versions by x.y.(z+1)

Change-Id: I4b054c308bd3f13c37b4a7905c2d8de1efe22220
Signed-off-by: jenkins-releng <jenkins-releng@opendaylight.org>


bug 7985 Using UncaughtExceptionHandler in TransactionInvokerImpl

when a particular transaction command fails , it is killing the only
worker processing the transaction commands queue.

when the exection is caught log the exception and resubmit the worker

if 10 commands are to be exectuded
while executing 5th command some exception happens
then the execution resumes from 6th command ignoring the 5th command.

Again if 6th command also fails because of some exception,
then execution resumes from 7th command onwards that way it wont result in
infinite loops.

This patch does not try to address transaction failures, only addressing any
exceptions in the commands.

for transaction chain failures, let us say all 10 commands are executed
successfully and submitted to transaction chain, now we wait for
success/failure of each such transaction asynchronously.
Now our pending tx list reads as 1,2,3,4,5,6,7,8,9,10

we may get success response for 1,2,3,6,7 and failure for 5th.
Then pending tx list reads as 4,5,8,9,10.

then the transactions will resume from 8th onwards. Note 4th is not resumed.
This behavior may not be desired as we are trying to resubmit 8th one which
already got executed.
If the transactions are idempotent then it should not be an issue at all to
reexecute the succeded ones.
Now if the 8th transaction is about deleting an entry ,
in its previous run it already deleted it, so this run will fail.

Now for the 4th one we may get a success callback or failure callback.
If we get failure callback , then again 8,9,10 will be reexecuted.

This patch is not addressing this scenario , can be taken up in later
patches with the right design.

Note the difference between command execution and transaction result.
This patch is addressing exceptions in command execution.

Change-Id: I93bba958784637a4752b860d23354506f0bd98bb
Signed-off-by: K.V Suneelu Verma <k.v.suneelu.verma@ericsson.com>


bug 7599 improving perf of ucast mac learning

when the device is connected all its macs are learnt in one update.
node.merge is called for each local ucast and remote ucast mac.
Now it is called only once.

when a mac update comes its physical locator need not be part of the
same update. If the locator is not available in the same update get
it from the cache ( which got filled from previous locators update msg).

Change-Id: Idbcf0b881b8ac90526f400c5d17d56a825f0a611
Signed-off-by: K.V Suneelu Verma <k.v.suneelu.verma@ericsson.com>


BUG 8280: TerminationPoint reconciliation fails

TerminationPointConfigReconciliationTask is incorrectly setting up
BridgeOperationalState to be empty. This ends up triggering reconciliation
even for ports that are already present in switch.

When we try to add ports that are already present we get constraint
violation in transaction and as per OVSDB RFC it causes all operations in
that transaction to abort. This means that the ports that actually needede
to be reconciled never get added back.

Change-Id: I48819f46b6419c15a8ba85541b06551b950a3055
Signed-off-by: Vishal Thapar <vishal.thapar@ericsson.com>


Add missing <name> fields for pom.xml files

This is used by autorelease scripts to automatically parse which project
is failing a build and report to the mailing list automatically. We need
names in the format:

    ODL :: <groupId> :: <artifactId>

This patch formats in the same format as found in the startup archetypes
patch found here: https://git.opendaylight.org/gerrit/52522

Change-Id: Ib3621b8a5f2aa1e0b50aa0968b0b728081993772
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>


BUG 6692 - Fix checkstyle problems not detected by the current version

This change is required for overall move to new Checkstyle version, see
https://git.opendaylight.org/gerrit/#/q/topic:bumpCheckstyle

Change-Id: If08990eaf8b849e6d826322eb1b7301996bc86c1
Signed-off-by: David <david.suarez.fuentes@ericsson.com>


bug 7781 update vlan bindings and tunnel ips

When physical switch update arrives write it in op datastore.

Change-Id: I46ca4f70b59961eeebbc17c0073fe40f852278b5
Signed-off-by: K.V Suneelu Verma <k.v.suneelu.verma@ericsson.com>


Don't ignore .gitignore

Change-Id: I0153a81206b7b61e806c43eebdd8169aec0c9e20
Signed-off-by: Stephen Kitt <skitt@redhat.com>


Bug 7619: correct the southbound artifact

The feature repository should be southbound-features4, not features4.

Change-Id: I03250768f0371929f61eb79f6eca32752e16afe7
Signed-off-by: Stephen Kitt <skitt@redhat.com>


Update .gitreview to stable/carbon

Change-Id: I6c605a5cc4c65c9798db90b46a2af8c5770ee8c4
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>