Updated git submodules
Project: ovsdb master 16b77b93af16583c819259a82438a6cd7257ce10

BUG 8280: TerminationPoint reconciliation fails

TerminationPointConfigReconciliationTask is incorrectly setting up
BridgeOperationalState to be empty. This ends up triggering reconciliation
even for ports that are already present in switch.

When we try to add ports that are already present we get constraint
violation in transaction and as per OVSDB RFC it causes all operations in
that transaction to abort. This means that the ports that actually needede
to be reconciled never get added back.

Change-Id: I48819f46b6419c15a8ba85541b06551b950a3055
Signed-off-by: Vishal Thapar <vishal.thapar@ericsson.com>