Updated git submodules
Project: infrautils master d4366e5d1359ab33a7e36de18a54b6c969515a86

Bug 8397: Move BP XML from ODL specific to standard location

and remove ODL specific xmlns:odl which are no used in this XML anyway

This is required so that BP errors are detected by SFT in the feature
including this bundle.  The underlying root cause of the problem was
that the respective feature (odl-infrautils-jobcoordinator) does not
depend on the ODL controller feature which installs the blueprint
handler thing, so SFT could never test for issues in this BP XML.

This requires the https://git.opendaylight.org/gerrit/#/c/54885/ which
just got merged in odlparent.  Without that, this would break J.C. BP.

PS: An alternative approach would be to, somehow, hard-code ALWAYS have
the ODL controller feature which installs the blueprint handler thing,
but that seems less modular to me.  Doing it like this allows to e.g.
write an IT in infrautils without relying on controller BP.

Change-Id: I53a1f046665ef62ed65b91c6dfce15070e2df54f
Signed-off-by: Michael Vorburger <vorburger@redhat.com>