Updated git submodules
Project: odlparent master 5a5c88b66af3e63532472ed02cfec99342369c07

Bug 8415 related: Make TestBundleDiag a lib

This is the alternative to c/56848 requested by Robert.

We're changing this so that we can consume bundles-test-lib in the
infrautils.ready implementation as a service for both Karaf v3 and v4.

As a (VERY nice!) side effect, this allows us to remove the duplicated
Java code we had between bundles-test and bundles4-test.  That was,
originally, intended to only be like it still is for a short while, but
with the Karaf 4 situation being what it is, this code duplication is
still there - and the respective code has started diverging a bit; so
unifying this in the new bundles-test-lib is great.

Change-Id: I0fd2c954fdc4a6c84505bef51729d15e4ba2888c
Signed-off-by: Michael Vorburger <vorburger@redhat.com>