Update patch set 4
Patch Set 4: Code-Review-1

(2 comments)

Not sure what's everybody's opinion on this, but I think it is helpful to state the 'domain' for which this *can* be used. Or perhaps better, to indicate that when federating against any authentication backend, none of this will work (I'm not even sure on what would the behavior be).

Patch-set: 4
Reviewer: Gerrit User 6419 <6419@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1