Updated git submodules
Project: releng/builder master 75b8d9a07513cd1a6323b72078df5ed45da3b345

Add a workaround for missing taglist.log

Since we fail the build now if any line in a script fails this is
causing validate-autorelease to fail unexpectedly. Temporarily put in
place a workaround on the problematic line so that it passes until we
put in a proper fix.

Change-Id: I31999f0f78d5ca360f20fa14536310924263c0a4
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>