Updated git submodules
Project: releng/builder master f95ac3188ea5414562f28476639f421a24473977

Merge "DO NOT fail autorelease if file scripts fail"

DO NOT fail autorelease if file scripts fail

There are 2 scripts which create a file in int/dist which are sure to
fail if int/dist is not included in autorelease. We noticed this because
of Nitrogen work.

Change-Id: I7a7123c551b420f838b281c810dc1711ef5c90d2
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>