Updated git submodules
Project: netvirt master 6ef2fa538dd911eef2f552920568f704e4e69361

Bug 8440: Traffic from remote OVS to LB Group

Traffic to extra-route from a remote switch was not going to the LB
group from the tunnel table.

This review makes sure that the tunnel table points to the LB group and
also makes sure that if the extra route could be reached through another
VM which is present in different switch then the tunnel table directly
points to only one of the VMs in order to avoid looping.

Change-Id: I0964991eef519610487e5454e39a643a4c04536b
Signed-off-by: Kency Kurian <kency.kurian@ericsson.com>