Updated git submodules
Project: netvirt master ddef7ee58970dfef4ef00c1933b748e412e87bb3

Fix NPE triggered after disabling SG on a port

When port-security is disabled on a Neutron port, removeAclInterfaceMap
is invoked, but the entry in aclInterfaceMap is not deleted as there is
a difference (i.e., securityGroups element) in the port info present in
aclInterfaceMap and the one passed as argument. Because of this, the stale
entries remain in the aclInterfaceMap and are causing NPE issues for
subsequent operations. This patch fixes this issue by matching on the
interfaceId and deleting the corresponding entry.

Change-Id: I2c98f82beab8043052c9444f40f02eb26e8c3945
Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>