Updated git submodules
Project: netvirt stable/carbon bdb38f669c48c01659b49bea48bb5511c9751d2f

Fix NPE triggered after disabling SG on a port

When port-security is disabled on a Neutron port, removeAclInterfaceMap
is invoked, but the entry in aclInterfaceMap is not deleted as there is
a difference (i.e., securityGroups element) in the port info present in
aclInterfaceMap and the one passed as argument. Because of this, the stale
entries remain in the aclInterfaceMap and are causing NPE issues for
subsequent operations. This patch fixes this issue by matching on the
interfaceId and deleting the corresponding entry.

Change-Id: I2c98f82beab8043052c9444f40f02eb26e8c3945
Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>