Updated git submodules
Project: netvirt stable/carbon 70391261d0be69edc11649d375faa7d69a945036

Bug 8179 - Data validation failed for path bgp/networks

Issue:

Data validation failed exception during withdrawing prefix from BGP.

Analysis:

The issue occured while trying to delete an element from the
bgp/networks which was not in the desired format. The DPNid is
used as rd in case of extra routes with internal VPN. The prefixes in
the internal VPN are not be advertised to the BGP. The condition for
this already exists(rd is the vpnName here). For extra routes, DPNIDs
are used as rds for ECMP and they should also be advertised. This was
already handled but there was no handling to avoid the withdrawal from
the BGP in case of the removal.

Fix:

Condition has been added to avoid calling withdrawPrefix() for the extra
routes removal in case of the internal VPN.

Note: Also some other errors in logging have been fixed.

Change-Id: Ic5dbcf951cd36b0a0fb9624f4899184bc4e305f0
Signed-off-by: Gobinath <gobinath@ericsson.com>
Signed-off-by: Vivekanandan Narasimhan <n.vivekanandan@ericsson.com>