Updated git submodules
Project: integration/test master bb48eb9c148e44f86391b1575ad8cf6fe17d1a54

Do not use different prefix in cluster tests

The different prefix was used as a workaround
for an issue that is already fixed.
On the other hand, prefix shard creation is not notified abou that,
so it is easiest to just the original prefix.

Change-Id: Ic83ce679e2a04fb05e9d897d2e04ffaafd05bdd5
Signed-off-by: Vratko Polak <vrpolak@cisco.com>