Update patch set 11
Patch Set 11:

OK. I basically give up. I can't for the life of me figure out what the RSTcheckBear is complaining about. It seems that it didn't like some, but not all, of my code-block directives when they specified bash as the language, but then had git commands as the first line. I fixed that between patch sets 9 and 10 by just removing bash as the language in those places, but I still can't fix this:

 10:03:40 Executing section reStructuredText...
 10:03:45 [33m[WARNING][0m[14:03:44] RSTcheckBear: This result has no patch attached.[0m
 10:03:45
 10:03:45 docs/gerrit.rst
 10:03:45 |  95| ···..·code-block::
 10:03:45 |    | [MAJOR] RSTcheckBear:
 10:03:45 |    | (ERROR/3) Error in "code-block" directive:

Patch-set: 11