Update patch set 4
Patch Set 4:

> > Also, Is there any way you know where I can test by shifting
 > > branches, because the ReadTheDocs version window doesn't display
 > > when local?
 > 
 > I do not know. I know ReadTheDocs' build system injects code into
 > your build and that's how it achieves the custom readthedocs
 > elements. Their build system is open source so you might be able to
 > get hints from looking at the readthedocs code.

Targeting to finish it before this weekend starts.

A good workaround is not to worry about the rtd version window (the one which gets displayed at the bottom of documentation) as it is, since I haven't been able to figure out a way to manually tell rtd what is the latest/stable and we don't really need anything to tweak in there? Probably we do it manually.

Patch-set: 4