Updated git submodules
Project: netvirt master 0e246207655296cda1173c490571e1abd69e41d5

Bug 8818 : BGP Manager. VPNv6 prefixes injected to QBGP

This issue is related to VPNv6 prefixes under the format prefix/length.
Indeed, the prefix/length format is not recognized by bgp manager when
analysing the prefix. Then it fails and considers it as IPv4.
Subsequently, an inconsistency is detected by QBGP, and it is rejected.
The fix consists in striping the length part so that the prefix can be
analysed correctly.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Change-Id: Ic396fe795eb6704c92b099c74631c96774034427