Updated git submodules
Project: releng/builder master 3e0433492ba64a4bad973384da83ca0bff54a526

Detect if staging has failed

The mvn command always exits 0 when using the staging plugin to stage a
Nexus repository. Due to this when failure occurs we do not currently
detect it. This patch parses the staging log file for any lines starting
with [ERROR] (typical of mvn errors) and exits the script with a failure
code if detected to make Jenkins fail the build.

Issue: RELENG-218
Change-Id: I5ddc56d88ca23baada1a8044ef9dc26345a583f9
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>