Updated git submodules
Project: odlparent master f2e0c03ebaa8c5ef3e9296d05ab7a72b774efdf7

Move karaf-util dependency

from bundles-test to features-test

The new karaf-util artifact introduced in
https://git.opendaylight.org/gerrit/#/c/61293/ should not be a
dependency of bundles-test but rather of features-test IMHO.

This is because bundles-test actually not have a compile time dependency
to the CustomBundleUrlStreamHandlerFactory from karaf-util; only
features-test (in PerRepoTestRunner).

However, bundles-test is also used in infrautils' ready.  Bumping to
odlparent 2.0.4 (which includes this) fell over with "...
odl-infrautils-ready ... Unable to resolve
org.opendaylight.odlparent.karaf-util...", because of the
<Embed-Dependency>bundles-test-lib in ready-impl.

We could add karaf-util to the <Embed-Dependency> of ready-impl to work
around this, at least temporarily to unblock, but the root cause is that
the dependency is wrong.

Change-Id: I031fbbc8fb9d8eebba3550ef0683b8363397d8a8
Signed-off-by: Michael Vorburger <vorburger@redhat.com>