Updated git submodules
Project: ovsdb master a7637041f73ce8438566702222e22030cde9e5b0

bug 8029 handle expired in transit entries

when ref count goes to zero , locator is kept in transit state.

if the device still has references to it from other sources like tunnels
the locator will not be deleted from device.

The transactions which wait for this locator will expire.

If the intransit state expires , query the device itself to get its
status and proceed with the transactions which depends on this locator.

added a timer job which triggers every 30sec and checks for expired
intransit keys and clears them.

Change-Id: I0022da4a3a2e748629bc031246f6e6c1509d27a4
Signed-off-by: K.V Suneelu Verma <k.v.suneelu.verma@ericsson.com>


bug 8029 added ref counts for physical locators.

remote mcast and ucast macs refer to physical locators

whenever a new mac is added or updated, update their refcount for that
tep.

When the ref count of the tep goes to zero , the device may delete that
tep after the current transaction.
In that case mark the locator as in transit.

The next transaction which depends on this locator will not use this
locator uuid and will wait for the locator in transit state to be
resolved in dependency queue.

Eventually controller will get locator delete update command from the
device and proceed the transactions which are waiting for this locator
state to be changed from in transit state.

Added onSuccess and onFailure methods to transaction command.
Clear the locator refs for the succeeded delete mcast macs immediately
upon the success of the device transaction.

Clearing the locator ref count immediately upon success of delete mcast
is needed so that the next transaction will not use the possibly stale locator.

Clear the intransit status of the failed commands upon failure.

Change-Id: I84d64f7deb20f801f14e524f88431807afac3a99
Signed-off-by: K.V Suneelu Verma <k.v.suneelu.verma@ericsson.com>