Updated git submodules
Project: releng/builder master 32d5f6631ae2a27333548e362fa378d601feeadc

Merge "Only use -all- job if it exists"

Only use -all- job if it exists

In an effort to reduce the number of CSIT jobs we are running
and creating massive backlogs in the jenkins queue, it was
decided in the integration meeting (8/24/2017) [0] that it would
be safe to remove the -only- version of CSIT jobs if there
was a corresponding -all- version of that same job.

the -all- job is loading the same project features as the
-only- job but will also include other OpenDaylight features
from other projects that have been determined to not be in
conflict with said features.

[0] https://meetings.opendaylight.org/opendaylight-integration/2017/integration/opendaylight-integration-integration.2017-08-24-16.05.html

Change-Id: I099c2e856fdcd2722b52fda60587dc0cd48af696
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>