Update patch set 5
Patch Set 5:

It's failing the coala check. If we look at the log here:
https://logs.opendaylight.org/releng/jenkins092/docs-tox-verify-oxygen/184/tox/tox-coala.log.gz

It says the commit message is too long:

Shortlog of the HEAD commit contains 62 character(s). This is 12 character(s) longer than the limit (62 > 50).

Here's a blog post that explains how to construct a good commit message and why we need this kind of check:

https://chris.beams.io/posts/git-commit/

Patch-set: 5
Label: Verified=0