Update patch set 4
Patch Set 4: Code-Review+1

Ok this script gives produces a milestone 0 status report that is auto-generated based on the submitted milestone 0 readouts. This provides an example of how we can produce the same table with the other milestone reports. I do not have any more cycles to spend on this at the moment so I hope it provides enough instructions so that someone can take it and run with it to complete the M1 - M4 status reports.

An, I hope this is sufficient to satisfy your request and that we can finally put this manual vs automation issue to rest for good.

Patch-set: 4