Updated git submodules
Project: genius master 6a2156671fc021da02cbdd54600b849a540793b2

Fix the listeners registration process

It is a bad practice to pass the 'this' out of a constructor, specially
in multi-threading environments, since the object is not yet fully
constructed. To fix that, we have added a register() method that will be
invoked by BluePrint just after the constructor; both the register() and
close() methods need to be declared as 'public' and not 'final',
otherwise BluePrint will not be able to call them.

The shutdown() method has been removed in favour of a simple close().

Change-Id: I7f63f8041b583fa266f49cebe9ad8aab672bae1f
Signed-off-by: David Suarez <david.suarez.fuentes@gmail.com>